pyproject.toml
[project]
name = "japanese-image-captioner-lite"
version = "0.1.0"
description = "Lightweight LLaVA-style Japanese image captioning training scripts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "accelerate>=0.30.0",
  "pillow>=10.0.0",
  "pyyaml>=6.0.0",
  "torch>=2.1.0",
  "tqdm>=4.66.0",
  "transformers>=4.41.0",
]

[project.optional-dependencies]
lora = ["peft>=0.11.0"]

[project.scripts]
jicl-train = "jicl.train:main"
jicl-generate = "jicl.generate:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
